+++ wp-content/plugins/wordpress-beta-tester/wp-beta-tester.php
@@ -13,7 +13,7 @@
  * Plugin URI:        https://wordpress.org/plugins/wordpress-beta-tester/
  * Description:       Allows you to easily upgrade to Beta releases.
  * Author:            Peter Westwood, Andy Fragen
- * Version:           2.0.4
+ * Version:           2.1.0
  * Network:           true
  * Author URI:        https://blog.ftwr.co.uk/
  * Text Domain:       wordpress-beta-tester
@@ -21,6 +21,8 @@
  * License:           GPL v2 or later
  * License URI:       https://www.opensource.org/licenses/GPL-2.0
  * GitHub Plugin URI: https://github.com/afragen/wordpress-beta-tester
+ * Requires at least: 3.1
+ * Requires PHP:      5.2.4
  */
 
 // Exit if called directly.